# spec/api.yml
openapi: 3.0.3
info:
  title: 診所預約系統 API
  description: |
    診所預約系統的 RESTful API 規格文件。

    每個 API 的 summary 欄位對應 Feature File 中的 step statement。
  version: 1.0.0
  contact:
    name: Clinic Booking System
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /patient/appointments:
    get:
      tags:
        - 病患預約
      summary: 病患查看預約列表
      description: |
        **Feature:** 病患查看預約.feature

        **Rules:**
        - 系統必須顯示預約日期與時段
        - 系統必須顯示醫師姓名
        - 系統必須顯示診療類型
        - 系統必須顯示預約狀態
        - 系統必須顯示預約建立時間
      operationId: getPatientAppointments
      parameters:
        - name: line_user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 顯示預約狀態如下
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'

  # ============================================
  # 病患建立預約.feature
  # ============================================
    post:
      tags:
        - 病患預約
      summary: 病患建立預約
      description: |
        **Feature:** 病患建立預約.feature

        **Rules:**
        - 病患不可在黑名單中
        - 時段剩餘分鐘數必須大於等於診療所需分鐘數
        - 病患當日不可有重複預約
        - 預約成功後必須扣除時段分鐘數
        - 預約成功後必須建立預約記錄
        - 預約成功後必須發送 LINE 通知
        - 不可選擇過去日期
        - 不可選擇超過 30 天後的日期
        - 每次預約僅能選擇一個診療項目

        **併發控制 (併發預約控制.feature):**
        - 使用 Row-Level Lock 鎖定時段記錄
        - 先取得鎖定者預約成功
        - 後取得鎖定者檢查時段餘量不足則失敗
        - 交易失敗時必須回滾所有變更
        - 時段餘量不足時返回錯誤訊息與替代選項
      operationId: createPatientAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: 預約記錄被建立，LINE 訊息發送至 {lineUserId}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: 操作失敗 - 時段剩餘分鐘數不足或其他驗證錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentErrorResponse'
        '403':
          description: 操作失敗 - 黑名單病患無法建立預約
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 操作失敗 - 病患當日已有預約
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 病患修改預約.feature
  # ============================================
  /patient/appointments/{appointmentId}:
    get:
      tags:
        - 病患預約
      summary: 病患查看預約詳情
      description: |
        **Feature:** 病患查看預約.feature

        取得單一預約的詳細資訊。
      operationId: getPatientAppointmentById
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 預約詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '404':
          description: 預約不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - 病患預約
      summary: 病患修改預約
      description: |
        **Feature:** 病患修改預約.feature

        **Rules:**
        - 僅可修改狀態為「已預約」的預約
        - 不可修改時段開始前 3 小時內的預約
        - 修改預約時必須立即釋放原時段分鐘數
        - 修改預約時必須檢查新時段餘量
        - 修改預約時必須扣除新時段分鐘數
        - 修改成功後必須發送 LINE 通知
      operationId: updatePatientAppointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
      responses:
        '200':
          description: 預約記錄被更新，時段 {slotId} 剩餘分鐘數為 {remainingMinutes}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: 操作失敗 - 新時段餘量不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 操作失敗 - 已取消/已完成狀態無法修改 或 時段開始前 3 小時內無法修改
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 病患取消預約.feature
  # ============================================
    delete:
      tags:
        - 病患預約
      summary: 病患取消預約
      description: |
        **Feature:** 病患取消預約.feature

        **Rules:**
        - 取消預約時必須更新預約狀態為「已取消」
        - 取消預約時必須釋放時段分鐘數
        - 取消成功後必須發送 LINE 通知
      operationId: cancelPatientAppointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 預約狀態為 "cancelled"，時段 {slotId} 剩餘分鐘數為 {remainingMinutes}，LINE 訊息發送至 {lineUserId}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '404':
          description: 預約不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # 管理員登入.feature
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    AppointmentStatus:
      type: string
      enum: [booked, checked_in, completed, no_show, cancelled]
      description: |
        預約狀態：
        - booked: 已預約
        - checked_in: 已報到
        - completed: 已完成
        - no_show: 未報到
        - cancelled: 已取消
    CreateAppointmentRequest:
      type: object
      required:
        - line_user_id
        - slot_id
        - treatment_id
      properties:
        line_user_id:
          type: string
        slot_id:
          type: string
        treatment_id:
          type: string
        patient_id:
          type: string
          description: 若已有病患資料則帶入
    UpdateAppointmentRequest:
      type: object
      properties:
        slot_id:
          type: string
        treatment_id:
          type: string
    AppointmentResponse:
      type: object
      properties:
        id:
          type: string
        patient_id:
          type: string
        slot_id:
          type: string
        treatment_id:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        status_label:
          type: string
          description: 狀態中文標籤
        appointment_date:
          type: string
          format: date
        slot_start_time:
          type: string
          example: "09:00"
        slot_end_time:
          type: string
          example: "09:30"
        doctor_name:
          type: string
        treatment_name:
          type: string
        created_at:
          type: string
          format: date-time
    AppointmentErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          example: 時段已滿
        alternative_slots:
          type: array
          description: 返回其他可用時段建議
          items:
            $ref: '#/components/schemas/AvailableSlotResponse'

    # ============================================
    # 管理員預約
    # ============================================
...
    AvailableSlotResponse:
      type: object
      properties:
        id:
          type: string
        schedule_id:
          type: string
        doctor_id:
          type: string
        doctor_name:
          type: string
        date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string
        remaining_minutes:
          type: integer

    # ============================================
    # 系統排程
    # ============================================
...